Get it actually working properly
It's essential to this method's accuracy that the inertia tensor is updated in between incrementing the body's orientation and correcting the velocity, so I've added in a new system that updates inverse inertia tensor in the SolverBodyInertia struct. Sadly that means storing an extra inertia tensor, which may be undesirable. I think it's work-aroundable by manually applying the delta_rotation when going between velocity and momentum, but that seems like it would be slower. I guess it would potentially only be necessary in some places and not others?