Fix issue with ray caster with max hits 1 (#869)
# Objective

Fixes #868 

## Solution

Make sure the ray cast result is pushed into the `RayHits` component.

<!--
Note: If your PR contains breaking changes relative to the latest release,
      remember to add a migration guide! See `migration-guides/README.md`.
-->

## Testing

I tested this locally against my own project, and it fixed the issue.

You can test this yourself by spawning a `RayCaster` with max_hits set to 1, and ensuring that the `RayHits` component gets updated with the result.

Are there unit tests or similar that I can add a test case to?