Prune manifolds without points (#879)
# Objective

Based on the doc comments for `ContactManifold` and the example character controller, it seems like an invariant that they should always have at least 1 point. In my project with a slightly modified version of the example 3d kinematic controller this was sometimes violated when standing still in certain spots with slightly rugged terrain. In the `kinematic_controller_collisions` function this leads to `deepest_penetration` being set to `Scalar::MIN`, leading to inf/NaN in `LinearVelocity` which then spreads throughout the system.

## Solution

Prune manifolds with no points after pruning the points.

## Testing

I can no longer get my project to crash with this PR, before I could easily trigger it within 20 seconds