Fix panic when despawning joints after despawning their bodies (#882)
# Objective

Fixes #835.

Removing a joint after its body has been removed results in a panic, as the island no longer exists.

## Solution

Return early if the island doesn't exist. This isn't the best fix (ideally we should remove the joint from the island as soon as the bodies are removed, but this is tricky currently), however it seems to fix the most imminent problem of crashes.

## Testing

Ran the example in #835. Crashed before, but not anymore.