Scale `skin_width` by `PhysicsLengthUnit` (#898)
# Objective

The `move_and_slide_2d` example sometimes gets stuck on ramps due to a skin width that is too small. This is because it's not accounting for the scale of the world.

## Solution

Scale the `skin_width` by `PhysicsLengthUnit`. This appears to fix the problem.

## Testing

Ran the `move_and_slide_2d` example.