Add system that asserts physics components are finite (#909)
# Objective

A lot of people (me included) have had to track down NaNs or infinities that causes avian internals to crash, and they can be difficult to track down because they cause crashes a lot later than they are introduced.

## Solution

Add a system when running with debug assertions that checks relevant components during `PhysicsSystems::First`.

For bugs in Avian internals we can potentially also schedule this system in other places.

### Questions
* I placed it in `PhysicsSchedulePlugin` because it feels more general purpose than the rest of `PhysicsDebugPlugin` and because it needs `self.schedule`, but I'm not sure if that's the best location for it?
* We could make the system use `assert!` instead of `debug_assert!` and make it runnable outside of `debug_assertions`?

## Testing

I modified one of the examples to introduce an infinity during Update and saw that it was caught.